src/requirements.txt
requests
pynacl
discord.py
